+++ src/main/java/com/project/model/Utilisateur.java
@@ -38,7 +38,7 @@ public class Utilisateur {
         return id;
     }
     
-    public void setId(Long id){
+    public void setId(Integer id){
         this.id = id;
     }
