+++ src/main/java/com/project/repository/UtilisateurRepository.java
@@ -5,7 +5,7 @@ import com.project.model.Utilisateur.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer> {
     List<Utilisateur> findByRole(Role role);
     Utilisateur findByEmail(String email); 
     long countByRole(Utilisateur.Role role);
